Fastapi MCP chatbot/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain-core
langchain-groq
langchain-community
langgraph
langchain-mcp-adapters
fastmcp
websockets
pydantic
requests
aiosqlite